Update MAKE RASPBERRY PI 4b.rtf
Clarify how to make a positive acknowledgment of a file delivery that decrypts and is understood, but only for successful file deliveries of an encrypted file. Once forced off encryption in either IP path's direction, plaintext is sent to communicate how to move to new IPs, no positive acknowledgment for plaintext file deliveries.